harness simplfications
resolve conflict: apply peer_provider rename to moved harness file
