+++ consensus/Cargo.toml
@@ -48,6 +48,7 @@ commonware-conformance.workspace = true
 commonware-consensus = { path = ".", features = ["mocks"] }
 commonware-resolver = { workspace = true, features = ["mocks"] }
 commonware-runtime = { workspace = true, features = ["test-utils"] }
+paste.workspace = true
 rand_chacha.workspace = true
 rstest.workspace = true
 tracing-subscriber.workspace = true
