+++ consensus/src/marshal/core/mailbox.rs
@@ -175,8 +175,8 @@ impl<V: Variant> Mailbox<V> {
             .flatten()
     }
 
-    /// A best-effort attempt to retrieve a given [Finalization] from local
-    /// storage. It is not an indication to go fetch the [Finalization] from the network.
+    /// A best-effort attempt to retrieve a given finalization from local
+    /// storage. It is not an indication to go fetch the finalization from the network.
     pub async fn get_finalization(
         &self,
         height: Height,
