+++ consensus/src/marshal/mocks/mod.rs
@@ -1,5 +1,4 @@
 pub mod application;
 pub mod block;
-pub mod harness;
 pub mod resolver;
 pub mod verifying;
