+++ examples/reshare/src/orchestrator/actor.rs
@@ -7,8 +7,8 @@ use crate::{
 };
 use commonware_consensus::{
     elector::Config as Elector,
-    marshal::{core::Mailbox as MarshalMailbox, standard::Standard},
-    simplex::{self, scheme, types::Context},
+    marshal::{core::Mailbox as MarshalMailbox, standard::StandardSimplex},
+    simplex::{self, scheme::Scheme as SimplexScheme, types::Context},
     types::{Epoch, Epocher, FixedEpocher, ViewDelta},
     CertifiableAutomaton, Relay,
 };
@@ -40,14 +40,14 @@ where
     H: Hasher,
     A: CertifiableAutomaton<Context = Context<H::Digest, C::PublicKey>, Digest = H::Digest>
         + Relay<Digest = H::Digest>,
-    S: Scheme,
+    S: Scheme + SimplexScheme<H::Digest>,
     L: Elector<S>,
     T: Strategy,
 {
     pub oracle: B,
     pub application: A,
     pub provider: Provider<S, C>,
-    pub marshal: MarshalMailbox<S, Standard<Block<H, C, V>>>,
+    pub marshal: MarshalMailbox<StandardSimplex<Block<H, C, V>, S>>,
     pub strategy: T,
 
     pub muxer_size: usize,
@@ -68,7 +68,7 @@ where
     H: Hasher,
     A: CertifiableAutomaton<Context = Context<H::Digest, C::PublicKey>, Digest = H::Digest>
         + Relay<Digest = H::Digest>,
-    S: Scheme,
+    S: Scheme + SimplexScheme<H::Digest>,
     L: Elector<S>,
     T: Strategy,
     Provider<S, C>: EpochProvider<Variant = V, PublicKey = C::PublicKey, Scheme = S>,
@@ -78,7 +78,7 @@ where
     application: A,
 
     oracle: B,
-    marshal: MarshalMailbox<S, Standard<Block<H, C, V>>>,
+    marshal: MarshalMailbox<StandardSimplex<Block<H, C, V>, S>>,
     provider: Provider<S, C>,
     strategy: T,
 
@@ -100,7 +100,7 @@ where
     H: Hasher,
     A: CertifiableAutomaton<Context = Context<H::Digest, C::PublicKey>, Digest = H::Digest>
         + Relay<Digest = H::Digest>,
-    S: scheme::Scheme<H::Digest, PublicKey = C::PublicKey>,
+    S: SimplexScheme<H::Digest, PublicKey = C::PublicKey>,
     L: Elector<S>,
     T: Strategy,
     Provider<S, C>: EpochProvider<Variant = V, PublicKey = C::PublicKey, Scheme = S>,
